feat: Task 116 NONROAD numerical-divergence triage (mo-490cm)
Task 115's end-to-end fidelity gate is dormant — it needs a captured
gfortran reference corpus and Task 117's run_simulation wiring, neither
of which exists yet. Task 116's actionable scope is therefore the
corpus-independent audit: the four instrumented-phase modules compared
against the pinned NONROAD2008a Fortran for divergences determinable
from source alone.

Findings:

- GETPOP / popeqp: the port carried equipment population as f64;
  getpop.f :211 reads it into valtmp, a real*4. Population values with
  more than ~7 significant digits would diverge under the 1e-9 budget.
  Fixed — PopulationRecord/SelectedPopulation `population` is now f32,
  parsed straight to f32; the fidelity adapter widens it like the
  other real*4 fields. Regression test pins the real*4 rounding.

- GRWFAC / factor: grwfac.f :244 divides by the un-clamped baseyearind,
  not the MINGRWIND-clamped tmpbaseyearind it computes for the
  divide-by-zero warning — so a zero base indicator gives the Fortran
  +/-Inf. The port already substitutes the clamp (the fix the warning
  text itself promises), keeping the factor finite. Kept that behavior
  but made it an explicit, documented divergence: the growth.rs
  comments claimed fidelity to grwfac.f :244-245 while silently
  diverging. Regression test pins the finite result.

- AGEDIST: audited, faithful — f32 throughout, Fortran evaluation
  order, left-to-right frcsum accumulation. No change.

- CLCEMS: audited — f32 throughout (f32::exp / f32::powf match the
  real*4 intrinsics). The only residual risk is Rust-libm vs gfortran
  last-bit differences in expf/powf, which is corpus-dependent and
  cannot be triaged until the gate activates.

characterization/nonroad-fidelity/README.md's known-divergences
section is populated with the triage record. The empirical,
per-fixture divergence triage remains gated on the gfortran corpus
plus Task 117.

+++ crates/moves-nonroad/src/population/pop.rs
@@ -64,8 +64,9 @@ pub struct SelectedPopulation {
     /// Technology/distribution code (10 chars, left-justified,
     /// upper-cased per `getpop.f` :141).
     pub tech_code: String,
-    /// Equipment population.
-    pub population: f64,
+    /// Equipment population. `f32` to match `getpop.f`'s `real*4`
+    /// `popeqp` — see `PopulationRecord::population` (Task 116).
+    pub population: f32,
 }
 
 /// Select population records for `scc` and `target_year` from a
@@ -159,7 +160,7 @@ fn slot_from(r: &PopulationRecord) -> SelectedPopulation {
 mod tests {
     use super::*;
 
-    fn rec(scc: &str, fips: &str, sub: &str, hp: f32, year: i32, pop: f64) -> PopulationRecord {
+    fn rec(scc: &str, fips: &str, sub: &str, hp: f32, year: i32, pop: f32) -> PopulationRecord {
         PopulationRecord {
             fips: fips.to_string(),
             subregion: sub.to_string(),
